Call assembler on resource assembly files
Instead of the janky way it was done before with the source code residing
within the SCons tool, SCons now detects dependencies on the assembly
files and they are just easier to maintain.

This also fixes #6 because SCons StaticObject() was compiling the
assembly files into ELFs, which was obviously WRONG for Windows. Now,
the correct assembler from MinGW is called, depending on the target
environment.